feat(promiseTracker): add `activationDelay` option
`activationDelay` `{Number}` - Number of milliseconds that an added
promise needs to be pending before this tracker is active.

Usage example: You have some http calls that sometimes return too
quickly for a loading spinner to look good. You only want to show the
tracker if a promise is pending for over 500ms.

You put `{activationDelay: 500}` in options.