Move angular from devDependencies to dependencies
Not sure what the minimum supported angular is, but it should be listed
in bower.json.

This is useful with libraries that analyze bower.json, e.g.
[main-bower-files](https://github.com/ck86/main-bower-files). (To use
main-bower-files with angular-promise-tracker, it's currently necessary to put in a manual override inside your own project's bower.json for angular-promise-tracker's dependency on angular: otherwise, it will load this package before angular.)